+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/metrics/TkmsMetricsTemplate.java
@@ -5,6 +5,7 @@ import com.transferwise.common.baseutils.meters.cache.TagsSet;
 import com.transferwise.common.context.TwContext;
 import com.transferwise.kafka.tkms.CompressionAlgorithm;
 import com.transferwise.kafka.tkms.api.TkmsShardPartition;
+import com.transferwise.kafka.tkms.config.TkmsProperties;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.Meter;
 import io.micrometer.core.instrument.Meter.Type;
@@ -33,32 +34,33 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
 
-  public static final String METRIC_LIBRARY_INFO = "tw_library_info";
-  public static final String PROXY_POLL = "tw_tkms_proxy_poll";
-  public static final String PROXY_CYCLE = "tw_tkms_proxy_cycle";
-  public static final String PROXY_CYCLE_PAUSE = "tw_tkms_proxy_cycle_pause";
-  public static final String PROXY_MESSAGE_SEND = "tw_tkms_proxy_message_send";
-  public static final String PROXY_KAFKA_MESSAGES_SEND = "tw_tkms_proxy_kafka_messages_send";
-  public static final String PROXY_MESSAGES_DELETION = "tw_tkms_proxy_messages_delete";
-  public static final String INTERFACE_MESSAGE_REGISTERED = "tw_tkms_interface_message_registration";
-  public static final String DAO_MESSAGE_INSERT = "tw_tkms_dao_message_insert";
-  public static final String DAO_MESSAGES_DELETION = "tw_tkms_dao_messages_delete";
-  public static final String DAO_POLL_FIRST_RESULT = "tw_tkms_dao_poll_first_result";
-  public static final String DAO_POLL_GET_CONNECTION = "tw_tkms_dao_poll_get_connection";
-  public static final String DAO_POLL_ALL_RESULTS = "tw_tkms_dao_poll_all_results";
-  public static final String DAO_POLL_ALL_RESULTS_COUNT = "tw_tkms_dao_poll_all_results_count";
-  public static final String DAO_INVALID_GENERATED_KEYS_COUNT = "tw_tkms_dao_insert_invalid_generated_keys_count";
-  public static final String DAO_ROWS_IN_TABLE_STATS = "tw_tkms_dao_rows_in_table_stats";
-  public static final String DAO_ROWS_IN_ENGINE_INDEPENDENT_TABLE_STATS = "tw_tkms_dao_rows_in_engine_independent_table_stats";
-  public static final String DAO_ROWS_IN_INDEX_STATS = "tw_tkms_dao_rows_in_index_stats";
-  public static final String DAO_APPROXIMATE_MESSAGES_COUNT = "tw_tkms_dao_approximate_messages_count";
-  public static final String STORED_MESSAGE_PARSING = "tw_tkms_stored_message_parsing";
-  public static final String MESSAGE_INSERT_TO_ACK = "tw_tkms_message_insert_to_ack";
-  public static final String DAO_COMPRESSION_RATIO_ACHIEVED = "tw_tkms_dao_serialization_compression_ratio";
-  public static final String DAO_ORIGINAL_SIZE_BYTES = "tw_tkms_dao_serialization_original_size_bytes";
-  public static final String DAO_SERIALIZED_SIZE_BYTES = "tw_tkms_dao_serialization_serialized_size_bytes";
-  public static final String DAO_EARLIEST_MESSAGE_ID = "tw_tkms_dao_earliest_message_id";
-  public static final String DAO_EARLIEST_MESSAGE_ID_COMMIT = "tw_tkms_proxy_poll_earliest_message_id_poll";
+  public static final String GAUGE_LIBRARY_INFO = "tw_library_info";
+  public static final String TIMER_PROXY_POLL = "tw_tkms_proxy_poll";
+  public static final String GAUGE_PROXY_POLL_IN_PROGRESS = "tw_tkms_proxy_poll_in_progress";
+  public static final String TIMER_PROXY_CYCLE = "tw_tkms_proxy_cycle";
+  public static final String TIMER_PROXY_CYCLE_PAUSE = "tw_tkms_proxy_cycle_pause";
+  public static final String COUNTER_PROXY_MESSAGE_SEND = "tw_tkms_proxy_message_send";
+  public static final String TIMER_PROXY_KAFKA_MESSAGES_SEND = "tw_tkms_proxy_kafka_messages_send";
+  public static final String TIMER_PROXY_MESSAGES_DELETION = "tw_tkms_proxy_messages_delete";
+  public static final String COUNTER_INTERFACE_MESSAGE_REGISTERED = "tw_tkms_interface_message_registration";
+  public static final String COUNTER_DAO_MESSAGE_INSERT = "tw_tkms_dao_message_insert";
+  public static final String COUNTER_DAO_MESSAGES_DELETION = "tw_tkms_dao_messages_delete";
+  public static final String COUNTER_DAO_MESSAGES_DELETION_BATCHES = "tw_tkms_dao_messages_delete_batches";
+  public static final String TIMER_DAO_POLL_FIRST_RESULT = "tw_tkms_dao_poll_first_result";
+  public static final String TIMER_DAO_POLL_GET_CONNECTION = "tw_tkms_dao_poll_get_connection";
+  public static final String TIMRE_DAO_POLL_ALL_RESULTS = "tw_tkms_dao_poll_all_results";
+  public static final String SUMMARY_DAO_POLL_ALL_RESULTS_COUNT = "tw_tkms_dao_poll_all_results_count";
+  public static final String COUNTER_DAO_INVALID_GENERATED_KEYS_COUNT = "tw_tkms_dao_insert_invalid_generated_keys_count";
+  public static final String GAUGE_DAO_ROWS_IN_TABLE_STATS = "tw_tkms_dao_rows_in_table_stats";
+  public static final String GAUGE_DAO_ROWS_IN_ENGINE_INDEPENDENT_TABLE_STATS = "tw_tkms_dao_rows_in_engine_independent_table_stats";
+  public static final String GAUGE_DAO_ROWS_IN_INDEX_STATS = "tw_tkms_dao_rows_in_index_stats";
+  public static final String GAUGE_DAO_APPROXIMATE_MESSAGES_COUNT = "tw_tkms_dao_approximate_messages_count";
+  public static final String TIMER_MESSAGE_INSERT_TO_ACK = "tw_tkms_message_insert_to_ack";
+  public static final String SUMMARY_DAO_COMPRESSION_RATIO_ACHIEVED = "tw_tkms_dao_serialization_compression_ratio";
+  public static final String COUNTER_DAO_ORIGINAL_SIZE_BYTES = "tw_tkms_dao_serialization_original_size_bytes";
+  public static final String COUNTER_DAO_SERIALIZED_SIZE_BYTES = "tw_tkms_dao_serialization_serialized_size_bytes";
+  public static final String GAUGE_DAO_EARLIEST_MESSAGE_ID = "tw_tkms_dao_earliest_message_id";
+  public static final String COUNTER_DAO_EARLIEST_MESSAGE_ID_COMMIT = "tw_tkms_dao_earliest_message_id_commit";
 
   public static final Tag NA_SHARD_TAG = Tag.of("shard", "N/A");
   public static final Tag NA_PARTITION_TAG = Tag.of("partition", "N/A");
@@ -68,23 +70,23 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   public static final Tag TAG_POLL_RESULTS_NOT_EMPTY = Tag.of("pollResult", "not_empty");
 
   private final IMeterCache meterCache;
+  private final TkmsProperties tkmsProperties;
 
   @PostConstruct
   public void init() {
     Map<String, double[]> slos = new HashMap<>();
     double[] defaultSlos = new double[]{1, 5, 25, 125, 625, 3125, 15625};
-    slos.put(PROXY_POLL, defaultSlos);
-    slos.put(PROXY_CYCLE, defaultSlos);
-    slos.put(PROXY_CYCLE_PAUSE, defaultSlos);
-    slos.put(DAO_POLL_FIRST_RESULT, defaultSlos);
-    slos.put(DAO_POLL_ALL_RESULTS, defaultSlos);
-    slos.put(DAO_POLL_GET_CONNECTION, defaultSlos);
-    slos.put(PROXY_KAFKA_MESSAGES_SEND, defaultSlos);
-    slos.put(PROXY_MESSAGES_DELETION, defaultSlos);
-    slos.put(STORED_MESSAGE_PARSING, defaultSlos);
-    slos.put(DAO_POLL_ALL_RESULTS_COUNT, defaultSlos);
-    slos.put(MESSAGE_INSERT_TO_ACK, new double[]{1, 5, 25, 125, 625, 3125, 15625});
-    slos.put(DAO_COMPRESSION_RATIO_ACHIEVED, new double[]{0.05, 0.1, 0.25, 0.5, 0.75, 1, 1.25, 2, 4});
+    slos.put(TIMER_PROXY_POLL, defaultSlos);
+    slos.put(TIMER_PROXY_CYCLE, defaultSlos);
+    slos.put(TIMER_PROXY_CYCLE_PAUSE, defaultSlos);
+    slos.put(TIMER_DAO_POLL_FIRST_RESULT, defaultSlos);
+    slos.put(TIMRE_DAO_POLL_ALL_RESULTS, defaultSlos);
+    slos.put(TIMER_DAO_POLL_GET_CONNECTION, defaultSlos);
+    slos.put(TIMER_PROXY_KAFKA_MESSAGES_SEND, defaultSlos);
+    slos.put(TIMER_PROXY_MESSAGES_DELETION, defaultSlos);
+    slos.put(SUMMARY_DAO_POLL_ALL_RESULTS_COUNT, defaultSlos);
+    slos.put(TIMER_MESSAGE_INSERT_TO_ACK, new double[]{1, 5, 25, 125, 625, 3125, 15625});
+    slos.put(SUMMARY_DAO_COMPRESSION_RATIO_ACHIEVED, new double[]{0.05, 0.1, 0.25, 0.5, 0.75, 1, 1.25, 2, 4});
 
     meterCache.getMeterRegistry().config().meterFilter(new MeterFilter() {
       @Override
@@ -111,9 +113,9 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyPoll(TkmsShardPartition shardPartition, int recordsCount, long startNanoTime) {
     meterCache
-        .timer(PROXY_POLL, TagsSet.of(
+        .timer(TIMER_PROXY_POLL, TagsSet.of(
             partitionTag(shardPartition),
-            pollResultTag(recordsCount > 0),
+            pollResultTag(recordsCount == 0),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
   }
@@ -121,7 +123,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyMessageSendSuccess(TkmsShardPartition shardPartition, String topic, Instant insertTime) {
     meterCache
-        .counter(PROXY_MESSAGE_SEND, TagsSet.of(
+        .counter(COUNTER_PROXY_MESSAGE_SEND, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition),
             successTag(true),
@@ -130,7 +132,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
 
     if (insertTime != null) {
       meterCache
-          .timer(MESSAGE_INSERT_TO_ACK, TagsSet.of(
+          .timer(TIMER_MESSAGE_INSERT_TO_ACK, TagsSet.of(
               partitionTag(shardPartition),
               shardTag(shardPartition),
               topicTag(topic)))
@@ -141,7 +143,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyMessageSendFailure(TkmsShardPartition shardPartition, String topic) {
     meterCache
-        .counter(PROXY_MESSAGE_SEND, TagsSet.of(
+        .counter(COUNTER_PROXY_MESSAGE_SEND, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition),
             successTag(false),
@@ -153,7 +155,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   public void recordMessageRegistering(String topic, TkmsShardPartition shardPartition) {
     TwContext currentContext = TwContext.current();
     meterCache
-        .counter(INTERFACE_MESSAGE_REGISTERED, TagsSet.of(
+        .counter(COUNTER_INTERFACE_MESSAGE_REGISTERED, TagsSet.of(
             entryPointGroupTag(currentContext),
             entryPointNameTag(currentContext),
             entryPointOwnerTag(currentContext),
@@ -167,7 +169,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   public void recordDaoMessageInsert(TkmsShardPartition shardPartition, String topic) {
     TwContext currentContext = TwContext.current();
     meterCache
-        .counter(DAO_MESSAGE_INSERT, TagsSet.of(
+        .counter(COUNTER_DAO_MESSAGE_INSERT, TagsSet.of(
             entryPointGroupTag(currentContext),
             entryPointNameTag(currentContext),
             entryPointOwnerTag(currentContext),
@@ -180,7 +182,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordDaoPollFirstResult(TkmsShardPartition shardPartition, long startNanoTime) {
     meterCache
-        .timer(DAO_POLL_FIRST_RESULT, TagsSet.of(
+        .timer(TIMER_DAO_POLL_FIRST_RESULT, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
@@ -189,12 +191,12 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordDaoPollAllResults(TkmsShardPartition shardPartition, int recordsCount, long startNanoTime) {
     meterCache
-        .timer(DAO_POLL_ALL_RESULTS, TagsSet.of(
+        .timer(TIMRE_DAO_POLL_ALL_RESULTS, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
     meterCache
-        .summary(DAO_POLL_ALL_RESULTS_COUNT, TagsSet.of(
+        .summary(SUMMARY_DAO_POLL_ALL_RESULTS_COUNT, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(recordsCount);
@@ -203,7 +205,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordDaoPollGetConnection(TkmsShardPartition shardPartition, long startNanoTime) {
     meterCache
-        .timer(DAO_POLL_GET_CONNECTION, TagsSet.of(
+        .timer(TIMER_DAO_POLL_GET_CONNECTION, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
@@ -212,9 +214,9 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyCycle(TkmsShardPartition shardPartition, int recordsCount, long startNanoTime) {
     meterCache
-        .timer(PROXY_CYCLE, TagsSet.of(
+        .timer(TIMER_PROXY_CYCLE, TagsSet.of(
             partitionTag(shardPartition),
-            pollResultTag(recordsCount > 0),
+            pollResultTag(recordsCount == 0),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
   }
@@ -222,7 +224,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyCyclePause(TkmsShardPartition shardPartition, long durationMs) {
     meterCache
-        .timer(PROXY_CYCLE_PAUSE, TagsSet.of(
+        .timer(TIMER_PROXY_CYCLE_PAUSE, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(durationMs, TimeUnit.MILLISECONDS);
@@ -231,7 +233,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyKafkaMessagesSend(TkmsShardPartition shardPartition, long startNanoTime) {
     meterCache
-        .timer(PROXY_KAFKA_MESSAGES_SEND, TagsSet.of(
+        .timer(TIMER_PROXY_KAFKA_MESSAGES_SEND, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
@@ -240,7 +242,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordProxyMessagesDeletion(TkmsShardPartition shardPartition, long startNanoTime) {
     meterCache
-        .timer(PROXY_MESSAGES_DELETION, TagsSet.of(
+        .timer(TIMER_PROXY_MESSAGES_DELETION, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(System.nanoTime() - startNanoTime, TimeUnit.NANOSECONDS);
@@ -254,20 +256,17 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordDaoMessagesDeletion(TkmsShardPartition shardPartition, int batchSize) {
     meterCache
-        .counter(DAO_MESSAGES_DELETION, TagsSet.of(
+        .counter(COUNTER_DAO_MESSAGES_DELETION_BATCHES, TagsSet.of(
             batchSizeTag(batchSize),
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .increment();
-  }
 
-  @Override
-  public void recordStoredMessageParsing(TkmsShardPartition shardPartition, long messageParsingStartNanoTime) {
     meterCache
-        .timer(STORED_MESSAGE_PARSING, TagsSet.of(
+        .counter(COUNTER_DAO_MESSAGES_DELETION, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
-        .record(System.nanoTime() - messageParsingStartNanoTime, TimeUnit.NANOSECONDS);
+        .increment(batchSize);
   }
 
   @Override
@@ -275,19 +274,19 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
       long serializedSizeBytes) {
     double ratio = (double) originalSizeBytes / serializedSizeBytes;
     meterCache
-        .summary(DAO_COMPRESSION_RATIO_ACHIEVED, TagsSet.of(
+        .summary(SUMMARY_DAO_COMPRESSION_RATIO_ACHIEVED, TagsSet.of(
             algorithmTag(algorithm),
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .record(ratio);
     meterCache
-        .counter(DAO_ORIGINAL_SIZE_BYTES, TagsSet.of(
+        .counter(COUNTER_DAO_ORIGINAL_SIZE_BYTES, TagsSet.of(
             algorithmTag(algorithm),
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .increment(originalSizeBytes);
     meterCache
-        .counter(DAO_SERIALIZED_SIZE_BYTES, TagsSet.of(
+        .counter(COUNTER_DAO_SERIALIZED_SIZE_BYTES, TagsSet.of(
             algorithmTag(algorithm),
             partitionTag(shardPartition),
             shardTag(shardPartition)))
@@ -297,7 +296,7 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
   @Override
   public void recordDaoInvalidGeneratedKeysCount(TkmsShardPartition shardPartition) {
     meterCache
-        .counter(DAO_INVALID_GENERATED_KEYS_COUNT, TagsSet.of(
+        .counter(COUNTER_DAO_INVALID_GENERATED_KEYS_COUNT, TagsSet.of(
             partitionTag(shardPartition),
             shardTag(shardPartition)))
         .increment();
@@ -310,31 +309,31 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
       version = "Unknown";
     }
 
-    Gauge.builder(METRIC_LIBRARY_INFO, () -> 1d).tags("version", version, "library", "tw-tkms")
+    Gauge.builder(GAUGE_LIBRARY_INFO, () -> 1d).tags("version", version, "library", "tw-tkms")
         .description("Provides metadata about the library, for example the version.")
         .register(meterCache.getMeterRegistry());
   }
 
   @Override
   public Object registerEarliestMessageId(TkmsShardPartition shardPartition, Supplier<Number> supplier) {
-    return registerGauge(DAO_EARLIEST_MESSAGE_ID, supplier, shardTag(shardPartition), partitionTag(shardPartition));
+    return registerGauge(GAUGE_DAO_EARLIEST_MESSAGE_ID, supplier, shardTag(shardPartition), partitionTag(shardPartition));
   }
 
   @Override
   public void registerRowsInTableStats(TkmsShardPartition sp, long rowsInTableStats) {
-    Gauge.builder(DAO_ROWS_IN_TABLE_STATS, () -> rowsInTableStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
+    Gauge.builder(GAUGE_DAO_ROWS_IN_TABLE_STATS, () -> rowsInTableStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
         .register(meterCache.getMeterRegistry());
   }
 
   @Override
   public void registerRowsInEngineIndependentTableStats(TkmsShardPartition sp, long rowsInTableStats) {
-    Gauge.builder(DAO_ROWS_IN_ENGINE_INDEPENDENT_TABLE_STATS, () -> rowsInTableStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
+    Gauge.builder(GAUGE_DAO_ROWS_IN_ENGINE_INDEPENDENT_TABLE_STATS, () -> rowsInTableStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
         .register(meterCache.getMeterRegistry());
   }
 
   @Override
   public void registerRowsInIndexStats(TkmsShardPartition sp, long rowsInIndexStats) {
-    Gauge.builder(DAO_ROWS_IN_INDEX_STATS, () -> rowsInIndexStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
+    Gauge.builder(GAUGE_DAO_ROWS_IN_INDEX_STATS, () -> rowsInIndexStats).tags(Tags.of(shardTag(sp), partitionTag(sp)))
         .register(meterCache.getMeterRegistry());
   }
 
@@ -384,6 +383,10 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
     return algorithm.getMicrometerTag();
   }
 
+  protected Tag databaseDialectTag(TkmsShardPartition sp) {
+    return Tag.of("databaseDialect", tkmsProperties.getDatabaseDialect(sp.getShard()).name());
+  }
+
   @Override
   public void unregisterMetric(Object rawMetricHandle) {
     if (rawMetricHandle == null) {
@@ -401,12 +404,17 @@ public class TkmsMetricsTemplate implements ITkmsMetricsTemplate {
 
   @Override
   public Object registerApproximateMessagesCount(TkmsShardPartition sp, Supplier<Number> supplier) {
-    return registerGauge(DAO_APPROXIMATE_MESSAGES_COUNT, supplier, shardTag(sp), partitionTag(sp));
+    return registerGauge(GAUGE_DAO_APPROXIMATE_MESSAGES_COUNT, supplier, shardTag(sp), partitionTag(sp));
+  }
+
+  @Override
+  public Object registerPollingInProgressGauge(TkmsShardPartition sp) {
+    return registerGauge(GAUGE_PROXY_POLL_IN_PROGRESS, () -> 1d, shardTag(sp), partitionTag(sp), databaseDialectTag(sp));
   }
 
   @Override
   public void registerEarliestMessageIdCommit(TkmsShardPartition shardPartition) {
-    meterCache.counter(DAO_EARLIEST_MESSAGE_ID_COMMIT, TagsSet.of(shardTag(shardPartition), partitionTag(shardPartition))).increment();
+    meterCache.counter(COUNTER_DAO_EARLIEST_MESSAGE_ID_COMMIT, TagsSet.of(shardTag(shardPartition), partitionTag(shardPartition))).increment();
   }
 
   protected MetricHandle registerGauge(String name, Supplier<Number> supplier, Tag... tags) {