+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/dao/PostgresTkmsDaoIntTest.java
@@ -15,9 +15,9 @@ class PostgresTkmsDaoIntTest extends TkmsDaoIntTest {
    */
   @Override
   protected void assertDeleteBucketsCounts() {
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "51").counter().count()).isEqualTo(19);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "11").counter().count()).isEqualTo(2);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "5").counter().count()).isEqualTo(2);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "51").counter().count()).isEqualTo(19);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "11").counter().count()).isEqualTo(2);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "5").counter().count()).isEqualTo(2);
   }
 
 }