+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/dao/TkmsDaoIntTest.java
@@ -57,9 +57,11 @@ class TkmsDaoIntTest extends BaseIntTest {
 
     tkmsDao.deleteMessages(TkmsShardPartition.of(0, 0), records);
 
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").counters()
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").counters()
         .stream().map(c -> c.count() * Integer.parseInt(c.getId().getTag("batchSize"))).reduce(0d, Double::sum))
         .isEqualTo(1001);
+
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").counter().count()).isEqualTo(1001);
 
     assertDeleteBucketsCounts();
 
@@ -67,10 +69,10 @@ class TkmsDaoIntTest extends BaseIntTest {
   }
 
   protected void assertDeleteBucketsCounts() {
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "256").counter().count()).isEqualTo(3);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "64").counter().count()).isEqualTo(3);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "16").counter().count()).isEqualTo(2);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "4").counter().count()).isEqualTo(2);
-    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete").tags("batchSize", "1").counter().count()).isEqualTo(1);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "256").counter().count()).isEqualTo(3);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "64").counter().count()).isEqualTo(3);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "16").counter().count()).isEqualTo(2);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "4").counter().count()).isEqualTo(2);
+    assertThat(meterRegistry.get("tw_tkms_dao_messages_delete_batches").tags("batchSize", "1").counter().count()).isEqualTo(1);
   }
 }