+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.2] - 2023-02-09
+
+### Fixed
+
+* Fixed bugs in some metrics.
+
+### Added
+
+* Added `tw_tkms_proxy_poll_in_progress` gauge with `databaseDialect` tag, for being able to distinguish Postgres and Maria services.
+* Added `tw_tkms_dao_messages_delete_batches`. `tw_tkms_dao_messages_delete` does not contain batches information anymore, it shows total deleted
+  records count.
+* Added `tw_tkms_dao_rows_in_engine_independent_table_stats`.
+
+### Removed
+
+* Removed `tw_tkms_stored_message_parsing` metric.
+
 ## [0.21.1] - 2023-02-06
 
 ### Changed